Revert "feat - #16 : Footer 컴포넌트 & Notice Board 컴포넌트"
