+++ src/app/(pages)/mypage/mypage.scss
@@ -40,11 +40,3 @@
   gap: 2rem;
   margin-top: 2rem;
 }
-
-.notice-section {
-  margin-top: 3rem;
-
-  @media (max-width: 768px) {
-    margin-top: 2rem;
-  }
-}
