--- src/components/blocks/NoticeBoard/NoticeBoard.types.ts
@@ -1,13 +1,0 @@
-export interface NoticeBoardProps {
-  title?: string
-  viewAllLink?: string
-  notices?: Notice[]
-  className?: string
-}
-
-export interface Notice {
-  id: number
-  title: string
-  date: string
-  href?: string
-}
