--- src/components/blocks/NoticeBoard/NoticeBoard_styles/NoticeBoard.scss
@@ -1,149 +1,0 @@
-.notice-section {
-  max-width: 132rem;
-  margin: 0 auto;
-}
-.notice-board {
-  background-color: var(--color-background-inverse);
-  padding: 6.4rem 2rem 11.2rem;
-  border-radius: 8px;
-  color: var(--color-text-primary);
-
-  &__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 3.5rem;
-  }
-
-  &__title {
-    font-size: 3.2rem;
-    font-weight: 400;
-    margin: 0;
-    color: var(--color-text-primary);
-  }
-
-  &__view-all {
-    font-size: 1.4rem;
-    color: var(--color-text-primary);
-    text-decoration: underline;
-    transition: color 0.2s ease;
-  }
-
-  &__content {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-  }
-
-  &__list {
-    display: flex;
-    width: calc(100% - 50rem);
-    ul {
-      list-style: none;
-      margin: 0 5rem 0 0;
-      padding: 0;
-      width: 100%;
-    }
-  }
-
-  &__item {
-    border-top: 1px solid var(--color-border-base);
-
-    &:last-child {
-      border-bottom: 1px solid var(--color-border-base);
-    }
-  }
-
-  &__link {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 2.4rem 0;
-    text-decoration: none;
-    color: white;
-    transition: background-color 0.2s ease;
-
-    // &:hover {
-    //   background-color: rgba(255, 255, 255, 0.05);
-    // }
-  }
-
-  &__notice-title {
-    font-size: 1.6rem;
-    font-weight: 400;
-    color: var(--color-text-primary);
-    flex: 1;
-    margin-right: 1rem;
-    line-height: 1.5;
-  }
-
-  &__notice-date {
-    font-size: 1.4rem;
-    color: var(--color-text-second);
-    white-space: nowrap;
-    font-weight: 400;
-  }
-
-  &__banner {
-    display: flex;
-    width: 50rem;
-    height: 29.3rem;
-    border-radius: 10px;
-    background: #BCE4F7;
-  }
-
-  // 반응형 디자인
-  @media (max-width: 768px) {
-    padding: 2rem 2rem 6.4rem;
-
-    &__header {
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-      margin-bottom: 1.6rem;
-      font-size: 2.1rem;
-    }
-
-    &__list {
-      width: 100%;
-      ul {
-        margin: 0;
-      }
-    }
-
-    &__title {
-      font-size: 2.1rem;
-    }
-
-    &__view-all {
-      font-size: 1.2rem;
-      align-self: center;
-    }
-
-    &__link {
-      flex-direction: column;
-      align-items: flex-start;
-      padding: 2.4rem 0;
-      gap: 0;
-    }
-
-    &__notice-title {
-      margin-right: 0;
-      margin-bottom: 0;
-      font-size: 1.6rem;
-      line-height: 1.5;
-      width: 100%;
-      // color: white;
-    }
-
-    &__notice-date {
-      font-size: 1.2rem;
-      margin-top: 0.8rem;
-    }
-
-    &__banner {
-      width: 100%;
-      margin-top: 3.2rem;
-    }
-  }
-}
