--- src/components/blocks/NoticeBoard/index.ts
@@ -1,2 +1,0 @@
-export { NoticeBoard } from './NoticeBoard'
-export type { NoticeBoardProps, Notice } from './NoticeBoard.types'
