+++ src/components/layouts/Footer/Footer_styles/Footer.scss
@@ -1,330 +1,79 @@
 .footer {
-  background-color: var(--color-background-inverse-toneup);
-  color: var(--color-text-inverse);
-  padding: 6.4rem 0 4rem;
-  @media (max-width: 1024px) {
-    padding: 6.4rem 2rem 4rem;
-  }
-
-  @media (max-width: 768px) {
-    padding: 2.4rem 0 0;
-  }
+  background-color: #212529;
+  color: white;
+  padding: 3rem 0 1rem;
 
   &__container {
-    max-width: 1440px;
+    max-width: 1200px;
     margin: 0 auto;
+    padding: 0 1rem;
   }
 
   &__content {
     display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 6.4rem;
-    padding-bottom: 4rem;
-    border-bottom: 1px solid var(--color-border-base);
-
-    @media (max-width: 1024px) {
-      grid-template-columns: repeat(3, 1fr);
-      gap: 2rem;
-    }
+    grid-template-columns: 1fr auto;
+    gap: 3rem;
+    margin-bottom: 2rem;
 
     @media (max-width: 768px) {
       grid-template-columns: 1fr;
-      gap: 0;
-      padding-bottom: 0;
+      gap: 2rem;
     }
   }
 
-  &__section {
-    @media (max-width: 768px) {
-      border-bottom: 1px solid var(--color-border-base);
-
-      &:last-child {
-        border-bottom: none;
-      }
-    }
+  &__brand {
+    max-width: 400px;
   }
 
-  &__section-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    background: none;
-    border: none;
-    cursor: pointer;
-    text-align: left;
-
-    @media (max-width: 768px) {
-      // cursor: default;
-      // pointer-events: none;
-      padding: 1.2rem 2rem;
-    }
-
-    &:focus {
-      outline: 2px solid #007bff;
-      outline-offset: 2px;
-    }
-  }
-
-  &__section-title {
-    font-size: 1.8rem;
+  &__logo {
+    font-size: 1.5rem;
     font-weight: 700;
-    margin: 0 0 1rem 0;
-    color: var(--color-text-primary);
-    line-height: 1.5;
-
-    @media (max-width: 768px) {
-      font-size: 1.6rem;
-      font-weight: 400;
-      margin: 0;
-    }
+    margin-bottom: 1rem;
+    color: white;
   }
 
-  &__section-icon {
-    display: none;
-    color: var(--color-icon-primary);
-    transition: transform 0.3s ease;
-
-    @media (max-width: 768px) {
-      display: block;
-    }
+  &__description {
+    color: #adb5bd;
+    line-height: 1.6;
   }
 
-  &__chevron {
-    transition: transform 0.3s ease;
-
-    &--open {
-      transform: rotate(180deg);
-    }
-  }
-
-  &__section-content {
-    @media (min-width: 769px) {
-      display: block;
-    }
-
-    @media (max-width: 768px) {
-      max-height: 0;
-      overflow: hidden;
-      transition: max-height 0.3s ease;
-
-      &--open {
-        max-height: 500px; // 충분한 높이로 설정
-      }
-    }
+  &__nav {
+    display: flex;
+    align-items: flex-start;
   }
 
   &__links {
     list-style: none;
+    display: flex;
+    gap: 2rem;
     margin: 0;
     padding: 0;
 
     @media (max-width: 768px) {
-      border-top: solid 1px var(--color-border-base);
-
-    }
-  }
-
-  &__link-item {
-    margin-bottom: 1.6rem;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-    @media (max-width: 768px) {
-      padding: 1.2rem 3.2rem;
-      margin: 0;
+      flex-direction: column;
+      gap: 1rem;
     }
   }
 
   &__link {
-    color: var(--color-text-base);
+    color: #adb5bd;
     text-decoration: none;
-    font-size: 1.6rem;
-    line-height: 1.5;
     transition: color 0.2s ease;
-    display: block;
-    font-weight: 400;
 
-    // &:hover {
-    //   color: #ccc;
-    // }
+    &:hover {
+      color: white;
+    }
   }
 
   &__bottom {
-    padding-top: 0.7rem;
-    text-align: left;
-    @media (max-width: 768px) {
-      padding: 0 2rem 2rem;
-    }
+    padding-top: 2rem;
+    border-top: 1px solid #495057;
+    text-align: center;
   }
 
   &__copyright {
-    color: var(--color-text-base);
-    font-size: 1.4rem;
+    color: #6c757d;
+    font-size: 0.875rem;
     margin: 0;
   }
-
-  /* Extras: social icons + app badges */
-  &__extras {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    gap: 2.4rem;
-    margin: 4rem 0;
-
-    @media (max-width: 768px) {
-      flex-direction: column-reverse;
-      align-items: flex-start;
-      padding: 0 2rem;
-    }
-  }
-
-  &__socials {
-    display: flex;
-    gap: 1.6rem;
-  }
-
-  &__social {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    background-color: var(--color-icon-base);
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    color: var(--color-icon-inverse-grey);
-    text-decoration: none;
-  }
-
-  &__badges {
-    display: flex;
-    gap: 2.4rem;
-    flex-wrap: wrap;
-  }
-
-  &__badge {
-    align-items: center;
-    height: 34px;
-    border-radius: .6rem;
-    background-color: var(--color-background-dark-alpha);
-    text-decoration: none;
-
-    &--gp {
-      background-color: #1f1f1f;
-    }
-
-    &--as {
-      background-color: #1f1f1f;
-    }
-  }
-
-  &__badge-text {
-    // font-size: 0.875rem;
-  }
-
-  /* Desktop company row + Mobile company info accordion */
-  &__company-desktop {
-    margin: 0.5rem 0 0.25rem;
-
-    @media (max-width: 768px) {
-      display: none;
-    }
-  }
-
-  /* Mobile company info accordion */
-  &__company-mobile {
-    display: none;
-    margin-top: 0.5rem;
-
-    @media (max-width: 768px) {
-      display: block;
-      border-top: 1px solid var(--color-border-base);
-      padding: 2rem;
-    }
-  }
-
-  &__company-toggle {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background: none;
-    color: var(--color-text-base);
-    border: none;
-    padding: 0.75rem 0;
-    text-align: left;
-  }
-
-
-  &__company-chevron {
-    color: var(--color-icon-primary);
-    transition: transform 0.2s ease;
-
-    &--open {
-      transform: rotate(180deg);
-    }
-  }
-
-  &__company-content {
-    max-height: 0;
-    overflow: hidden;
-    transition: max-height 0.25s ease;
-
-    &--open {
-      max-height: 200px;
-    }
-  }
-
-  &__company-info {
-    display: flex;
-    justify-self: start;
-    flex-wrap: wrap;
-    &-text {
-      display: flex;
-      color: var(--color-text-base);
-      font-size: 1.4rem;
-      line-height: 2.2rem;
-      gap: 0.8rem;
-      margin-right: 0.8rem;
-      strong {
-        margin-right: 1.6rem;
-      }
-
-      &::after {
-        display: flex;
-        content: '|';
-        color: #D9DBE1;
-      }
-      &:nth-child(3n) {
-        margin-right: 45%;
-        &::after {
-          display: none;
-        }
-      }
-    }
-    @media (max-width: 1024px) {
-      &-text {
-        &:nth-child(3n) {
-          margin-right: 0.8rem;
-        }
-      }
-    }
-
-    @media (max-width: 768px) {
-
-      &-text {
-        &:nth-child(3n) {
-          margin-right: 0;
-        }
-        &:last-child {
-          &::after {
-            display: none;
-          }
-        }
-      }
-    }
-
-  }
-
 }
