+++ src/components/layouts/Header/Header_styles/Header.scss
@@ -1,205 +1,55 @@
-@mixin button-reset {
-  background: none;
-  border: none;
-  padding: 0;
-  margin: 0;
-  font-family: inherit;
-  cursor: pointer;
-  appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  text-align: inherit;
-  text-decoration: none;
-  line-height: inherit;
-
-  &::-moz-focus-inner {
-    border: 0;
-    padding: 0;
-  }
-}
-
-.kal-header-wrapper {
+.header {
+  padding: 1rem 0;
+  background-color: var(--color-background-inverse);
   border-bottom: 1px solid var(--color-border-base);
-  background-color: var(--color-background-inverse-toneup);
-}
+  transition: background-color 0.3s ease, border-color 0.3s ease;
 
-.kal-header {
-  $utils-list-height: 4rem;
-
-  display: flex;
-  max-width: 144rem;
-  margin: 0 auto;
-  padding-top: $utils-list-height;
-
-  &__logo {
-    order: 1;
-    flex-shrink: 0;
+  &__container {
+    max-width: 1200px;
+    margin: 0 auto;
     display: flex;
+    justify-content: space-between;
     align-items: center;
-    justify-content: center;
-    color: --color-icon-primary;
+    padding: 0 1rem;
+    gap: 2rem;
 
-    &-link,
-    &-sky-link {
-      display: flex;
-      color: var(--color-icon-primary);
+    @media (max-width: 768px) {
+      flex-direction: column;
+      gap: 1rem;
     }
   }
 
-  &__utils {
-    order: 3;
-    flex-shrink: 0;
-    display: flex;
-    position: relative;
-    padding-top: 2rem;
-
-    &-list {
-      list-style: none;
-      margin: 0;
-      padding: 0;
-
-      display: flex;
-      align-items: center;
-      justify-content: flex-start;
-      gap: 0 2.4rem;
-      position: absolute;
-      top: -$utils-list-height;
-      right: 0;
-      width: max-content;
-      height: $utils-list-height;
-      padding: 1.2rem 0 0.4rem;
-    }
-
-    &-item {
-      display: flex;
-      flex-shrink: 0;
-    }
-
-    &-link,
-    &-button {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0 0.4rem;
-      font-size: 1.4rem;
-      color: var(--color-text-primary);
-    }
-
-    &-link {
-      text-decoration: none;
-    }
-
-    &-button {
-      @include button-reset;
-    }
-
-    .kal-header__actions {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0 1.6rem;
-    }
-
-    .kal-header-search__button {
-      @include button-reset;
-
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 0 1.6rem;
-      height: 4rem;
-      border-radius: 0.8rem;
-      border: 1px solid var(--color-border-base);
-      padding: 0 1.5rem;
-      background-color: var(--color-background-inverse_toneup);
-      color: var(--color-text-second);
-      font-size: 1.4rem;
-
-      &-icon {
-        color: var(--color-icon-primary);
-      }
-    }
-
-    .kal-header-auth-link {
-      text-decoration: none;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      min-width: 11.2rem;
-      height: 4rem;
-      border-radius: 999.9rem;
-      border: 2px solid var(--color-border-primary-darkble);
-      font-weight: 700;
-      font-size: 1.6rem;
-      color: var(--color-text-primary);
-    }
+  &__logo-link {
+    text-decoration: none;
+    color: inherit;
   }
-
-  &__nav-wrapper {
-    order: 2;
-    flex: 1;
-    display: flex;
-    padding: 0.9rem 3.2rem 0 10rem;
-  }
-
-  &__nav {
-    &-list {
-      list-style: none;
-      margin: 0;
-      padding: 0;
-
-      display: flex;
-      justify-content: flex-start;
-      flex-wrap: wrap;
-      column-gap: 2.4rem;
-    }
-
-    &-link {
-      text-decoration: none;
-
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 1.7rem 2rem 2.1rem;
-      font-size: 2.1rem;
-      color: var(--color-text-primary);
-    }
-  }
-}
-
-.kal-mobile-header {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  padding: 1.2rem 2rem;
 
   &__logo {
-    margin-right: auto;
-    padding: 0.3rem 0;
+    font-size: 1.5rem;
+    font-weight: 700;
+    color: var(--color-text-base);
+    margin: 0;
+    transition: color 0.3s ease;
 
-    &-link {
-      display: flex;
-      color: var(--color-icon-primary);
+    &:hover {
+      color: var(--color-text-primary);
     }
   }
 
   &__actions {
     display: flex;
+    gap: 1rem;
     align-items: center;
-    gap: 0 1.2rem;
+
+    @media (max-width: 768px) {
+      justify-content: center;
+    }
   }
 
-  &__menu-button {
-    @include button-reset;
-
-    width: 2.8rem;
-    height: 2.8rem;
-    color: var(--color-icon-primary);
-  }
-
-  &__auth-link {
-    text-decoration: underline;
-    font-size: 1.4rem;
-    color: var(--color-text-primary);
+  &__username {
+    font-size: 0.875rem;
+    color: var(--color-text-second);
+    white-space: nowrap;
+    transition: color 0.3s ease;
   }
 }
