--- src/components/svgs/AppStoreDownIcon.tsx
@@ -1,57 +1,0 @@
-import React from 'react';
-
-interface AppStoreDownIconProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const AppStoreDownIcon: React.FC<AppStoreDownIconProps> = ({
-  className = '',
-  width = 118,
-  height = 34,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      fill="none"
-      viewBox="0 0 118 34"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path fillRule="evenodd" clipRule="evenodd" d="M25.9124 16.8134C25.8862 13.9448 28.2509 12.55 28.3591 12.4852C27.0205 10.5243 24.9438 10.2562 24.2143 10.2352C22.4709 10.053 20.7815 11.2831 19.8932 11.2831C18.9858 11.2831 17.6193 10.2545 16.1464 10.2843C14.2494 10.3141 12.4745 11.4163 11.5007 13.1265C9.49375 16.6189 10.9911 21.7479 12.916 24.5708C13.8776 25.9534 15.0015 27.4972 16.4744 27.4411C17.9142 27.3815 18.4535 26.5194 20.1917 26.5194C21.9124 26.5194 22.4168 27.4411 23.9176 27.4061C25.4638 27.3815 26.4377 26.0182 27.3661 24.6234C28.4795 23.0411 28.9263 21.4798 28.942 21.3992C28.9071 21.3869 25.942 20.2514 25.9124 16.8134Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M23.0771 8.37827C23.8502 7.40574 24.3807 6.08275 24.2341 4.74048C23.1137 4.78954 21.7106 5.5185 20.9043 6.47001C20.1906 7.30936 19.5501 8.68317 19.7159 9.97462C20.9759 10.0692 22.2673 9.33678 23.0771 8.37827Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M43.3081 22.3047L42.3116 19.2136C42.2052 18.8982 42.0062 18.1535 41.7148 16.9829H41.6799C41.5629 17.4858 41.378 18.2306 41.1197 19.2136L40.1406 22.3047H43.3081ZM46.7551 27.2424H44.7429L43.6399 23.7641H39.8075L38.7587 27.2424H36.7988L40.5946 15.4021H42.9401L46.7551 27.2424Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M54.5954 22.9382C54.5954 22.1076 54.4104 21.4224 54.0369 20.8827C53.6268 20.3202 53.0788 20.0398 52.3912 20.0398C51.9253 20.0398 51.5029 20.1958 51.1225 20.5042C50.7438 20.8161 50.4942 21.2226 50.379 21.7273C50.3215 21.9604 50.29 22.1549 50.29 22.3056V23.7284C50.29 24.3505 50.482 24.8745 50.859 25.302C51.2394 25.7296 51.7316 25.9434 52.3389 25.9434C53.0492 25.9434 53.6059 25.6682 54.0003 25.118C54.3982 24.5678 54.5954 23.8406 54.5954 22.9382ZM56.5219 22.8678C56.5219 24.3205 56.1293 25.4682 55.3492 26.3128C54.6476 27.0628 53.7785 27.4361 52.7419 27.4361C51.6215 27.4361 50.8152 27.033 50.3266 26.2235H50.2899V30.7216H48.4016V21.515C48.4016 20.6021 48.3772 19.6663 48.3301 18.7043H49.9932L50.0979 20.0571H50.1328C50.7646 19.0373 51.721 18.5273 53.0037 18.5273C54.0071 18.5273 54.8448 18.9269 55.5132 19.7224C56.1851 20.5197 56.5219 21.5693 56.5219 22.8678Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M64.3766 22.9382C64.3766 22.1076 64.1899 21.4224 63.8164 20.8827C63.4063 20.3202 62.8601 20.0398 62.1707 20.0398C61.7048 20.0398 61.2824 20.1958 60.902 20.5042C60.5233 20.8161 60.2755 21.2226 60.1585 21.7273C60.101 21.9604 60.0713 22.1549 60.0713 22.3056V23.7284C60.0713 24.3505 60.2615 24.8745 60.6385 25.302C61.0172 25.7296 61.5111 25.9434 62.1184 25.9434C62.8304 25.9434 63.3854 25.6682 63.7798 25.118C64.1777 24.5678 64.3766 23.8406 64.3766 22.9382ZM66.3014 22.8678C66.3014 24.3205 65.9105 25.4682 65.1287 26.3128C64.4271 27.0628 63.558 27.4361 62.5214 27.4361C61.401 27.4361 60.5965 27.033 60.1061 26.2235H60.0712V30.7216H58.1811V21.515C58.1811 20.6021 58.1584 19.6663 58.1113 18.7043H59.7727L59.8792 20.0571H59.9141C60.5441 19.0373 61.5005 18.5273 62.7832 18.5273C63.7849 18.5273 64.6243 18.9269 65.2945 19.7224C65.9646 20.5197 66.3014 21.5693 66.3014 22.8678Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M77.2376 23.9231C77.2376 24.9289 76.8904 25.749 76.1905 26.3815C75.4227 27.0737 74.3529 27.4189 72.9794 27.4189C71.7089 27.4189 70.6915 27.1718 69.9219 26.6794L70.3599 25.0989C71.1889 25.6035 72.0999 25.8559 73.0911 25.8559C73.8014 25.8559 74.3564 25.6929 74.7543 25.3705C75.1504 25.0498 75.3476 24.617 75.3476 24.0773C75.3476 23.5954 75.1853 23.1906 74.859 22.8629C74.5326 22.5352 73.9899 22.2286 73.2307 21.9465C71.1662 21.1737 70.1348 20.0399 70.1348 18.5522C70.1348 17.578 70.496 16.7789 71.2203 16.1568C71.941 15.5365 72.9044 15.2246 74.1068 15.2246C75.1801 15.2246 76.0719 15.4121 76.7822 15.7853L76.311 17.3326C75.646 16.9699 74.8939 16.7894 74.0544 16.7894C73.3895 16.7894 72.8695 16.9524 72.496 17.2801C72.1836 17.5727 72.0248 17.9302 72.0248 18.3525C72.0248 18.8203 72.2046 19.2076 72.5658 19.5108C72.8817 19.7946 73.4524 20.096 74.2831 20.4255C75.297 20.8355 76.0422 21.3156 76.5221 21.8659C76.9986 22.4178 77.2376 23.103 77.2376 23.9231Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M83.4853 20.128H81.4033V24.274C81.4033 25.3289 81.7698 25.8563 82.5046 25.8563C82.8431 25.8563 83.1241 25.8265 83.3457 25.7687L83.3963 27.2091C83.0246 27.3475 82.5342 27.4194 81.9269 27.4194C81.1817 27.4194 80.5988 27.1898 80.1765 26.7342C79.7576 26.2769 79.5465 25.5093 79.5465 24.4317V20.128H78.3057V18.7051H79.5465V17.1403L81.4033 16.5796V18.7051H83.4853V20.128Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M90.9217 22.9654C90.9217 22.1786 90.7542 21.5022 90.4156 20.9397C90.0177 20.2563 89.4523 19.9164 88.7176 19.9164C87.9584 19.9164 87.3825 20.2563 86.9864 20.9397C86.6478 21.5022 86.4785 22.1891 86.4785 23.0022C86.4785 23.789 86.6478 24.4618 86.9864 25.0261C87.3947 25.7077 87.9671 26.0494 88.7019 26.0494C89.4244 26.0494 89.9898 25.7025 90.3982 25.0086C90.7472 24.4338 90.9217 23.7504 90.9217 22.9654ZM92.8827 22.9038C92.8827 24.2163 92.5092 25.2922 91.7623 26.1368C90.9804 27.0042 89.9403 27.437 88.6471 27.437C87.3976 27.437 86.4046 27.02 85.6646 26.1894C84.9229 25.3571 84.5547 24.3092 84.5547 23.044C84.5547 21.721 84.9334 20.6363 85.6995 19.7935C86.4622 18.9506 87.4918 18.5283 88.7867 18.5283C90.0345 18.5283 91.038 18.9436 91.7972 19.7777C92.5214 20.5855 92.8827 21.6264 92.8827 22.9038Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M99.025 20.3725C98.8383 20.3375 98.6376 20.32 98.4282 20.32C97.7633 20.32 97.2502 20.5723 96.8872 21.0752C96.5748 21.5203 96.416 22.0845 96.416 22.7627V27.2433H94.5277L94.5452 21.3924C94.5452 20.4093 94.5208 19.5121 94.4736 18.7043H96.1193L96.1891 20.3375H96.2398C96.4404 19.7767 96.7546 19.3246 97.1856 18.9847C97.608 18.6798 98.0617 18.5273 98.5521 18.5273C98.7266 18.5273 98.8837 18.5396 99.025 18.5641V20.3725Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M105.672 22.0767C105.685 21.516 105.563 21.0289 105.306 20.6188C104.98 20.0914 104.475 19.8285 103.8 19.8285C103.18 19.8285 102.679 20.0844 102.294 20.6013C101.981 21.0131 101.793 21.5037 101.735 22.0767H105.672ZM107.474 22.5693C107.474 22.9093 107.453 23.1967 107.406 23.4297H101.736C101.759 24.2743 102.033 24.9174 102.56 25.3643C103.036 25.762 103.656 25.96 104.413 25.96C105.253 25.96 106.017 25.8269 106.706 25.5553L107.003 26.873C106.199 27.2252 105.248 27.4004 104.152 27.4004C102.832 27.4004 101.797 27.0114 101.045 26.2316C100.293 25.4536 99.916 24.4075 99.916 23.0968C99.916 21.8088 100.265 20.7347 100.967 19.8813C101.701 18.9683 102.693 18.511 103.94 18.511C105.166 18.511 106.094 18.9683 106.724 19.8813C107.225 20.6067 107.474 21.5039 107.474 22.5693Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M41.7028 8.14041C41.7028 7.46052 41.523 6.94008 41.1652 6.57736C40.8075 6.21463 40.2857 6.03414 39.5963 6.03414C39.3031 6.03414 39.0536 6.05517 38.8477 6.09547V10.4447C38.9628 10.4622 39.1723 10.471 39.4777 10.471C40.188 10.471 40.7342 10.273 41.1216 9.8752C41.509 9.47918 41.7028 8.90091 41.7028 8.14041ZM42.6842 8.11421C42.6842 9.16209 42.3719 9.95063 41.7453 10.4798C41.1659 10.9687 40.3457 11.214 39.2794 11.214C38.7524 11.214 38.2986 11.1895 37.9199 11.1457V5.42091C38.4138 5.34031 38.9461 5.30176 39.5202 5.30176C40.5342 5.30176 41.3003 5.52255 41.8152 5.96588C42.3946 6.46529 42.6842 7.18198 42.6842 8.11421Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M46.9117 9.05993C46.9117 8.67267 46.8297 8.34149 46.6622 8.06637C46.4685 7.72993 46.191 7.56346 45.8297 7.56346C45.4563 7.56346 45.1735 7.72993 44.9781 8.06637C44.8123 8.34149 44.7285 8.67968 44.7285 9.07745C44.7285 9.46471 44.8123 9.7959 44.9781 10.0728C45.1788 10.4092 45.4597 10.5757 45.821 10.5757C46.1753 10.5757 46.4527 10.4057 46.6552 10.064C46.8262 9.78188 46.9117 9.44719 46.9117 9.05993ZM47.8776 9.02992C47.8776 9.67477 47.6926 10.204 47.3261 10.6193C46.9422 11.0451 46.4326 11.2571 45.7956 11.2571C45.183 11.2571 44.6944 11.0538 44.3297 10.6438C43.9667 10.2355 43.7852 9.72033 43.7852 9.09826C43.7852 8.44815 43.9701 7.9172 44.3471 7.50191C44.7223 7.08836 45.2284 6.88159 45.8637 6.88159C46.478 6.88159 46.9701 7.08486 47.3436 7.49315C47.6978 7.89092 47.8776 8.40259 47.8776 9.02992Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M54.8374 6.9668L53.5303 11.1636H52.6804L52.1376 9.33942C51.9998 8.88732 51.8898 8.43523 51.8026 7.98489H51.7869C51.7048 8.44574 51.5949 8.89784 51.4518 9.33942L50.8741 11.1636H50.0138L48.7852 6.9668H49.7398L50.2127 8.96092C50.3261 9.43229 50.4221 9.88088 50.4972 10.3067H50.5129C50.5827 9.95623 50.6961 9.50939 50.8584 8.96968L51.4518 6.9668H52.2057L52.7746 8.92588C52.9107 9.40426 53.0242 9.86336 53.1097 10.3067H53.1341C53.1987 9.87563 53.2929 9.41477 53.4203 8.92588L53.9264 6.9668H54.8374Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M59.6526 11.1625H58.7242V8.76006C58.7242 8.01884 58.4449 7.6491 57.8812 7.6491C57.6072 7.6491 57.3839 7.75074 57.2128 7.954C57.0401 8.15727 56.9528 8.39734 56.9528 8.67245V11.1625H56.0244V8.16603C56.0244 7.79805 56.0139 7.39852 55.9912 6.9657H56.808L56.8516 7.62107H56.876C56.986 7.4178 57.1448 7.25133 57.3594 7.11465C57.609 6.96045 57.89 6.88159 58.2006 6.88159C58.5915 6.88159 58.9144 7.00776 59.1727 7.26009C59.492 7.57025 59.6526 8.03461 59.6526 8.64967V11.1625Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M61.2872 11.1628H62.2139V5.04199H61.2872V11.1628Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M66.7174 9.05993C66.7174 8.67267 66.6336 8.34149 66.4678 8.06637C66.2741 7.72993 65.9966 7.56346 65.6354 7.56346C65.2619 7.56346 64.9775 7.72993 64.7855 8.06637C64.6179 8.34149 64.5342 8.67968 64.5342 9.07745C64.5342 9.46471 64.6179 9.7959 64.7855 10.0728C64.9844 10.4092 65.2654 10.5757 65.6267 10.5757C65.9809 10.5757 66.2584 10.4057 66.4609 10.064C66.6319 9.78188 66.7174 9.44719 66.7174 9.05993ZM67.6833 9.02992C67.6833 9.67477 67.4983 10.204 67.1318 10.6193C66.7478 11.0451 66.2383 11.2571 65.6013 11.2571C64.987 11.2571 64.5001 11.0538 64.1371 10.6438C63.7723 10.2355 63.5908 9.72033 63.5908 9.09826C63.5908 8.44815 63.7758 7.9172 64.1528 7.50191C64.528 7.08836 65.0341 6.88159 65.6693 6.88159C66.2836 6.88159 66.7758 7.08486 67.1492 7.49315C67.5035 7.89092 67.6833 8.40259 67.6833 9.02992Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M71.2146 9.72017V9.07357C70.1902 9.05605 69.6771 9.33817 69.6771 9.91818C69.6771 10.1372 69.7347 10.3019 69.8568 10.4123C69.9755 10.521 70.1273 10.5753 70.3088 10.5753C70.513 10.5753 70.7032 10.5105 70.8778 10.3808C71.0505 10.2511 71.1587 10.0864 71.1989 9.88489C71.2093 9.83757 71.2146 9.78325 71.2146 9.72017ZM72.1798 11.1625H71.3439L71.2741 10.6788H71.2497C70.9634 11.0643 70.5568 11.2571 70.028 11.2571C69.6336 11.2571 69.316 11.1309 69.0734 10.8769C68.8553 10.6473 68.7471 10.3617 68.7471 10.0235C68.7471 9.51005 68.96 9.11929 69.3875 8.84943C69.8151 8.57958 70.4172 8.44465 71.1903 8.45166V8.3728C71.1903 7.82083 70.9006 7.54572 70.323 7.54572C69.9111 7.54572 69.5481 7.6491 69.234 7.85763L69.0455 7.24432C69.4329 7.00075 69.9128 6.88159 70.4783 6.88159C71.5708 6.88159 72.1187 7.45985 72.1187 8.61638V10.1619C72.1187 10.5807 72.1379 10.9154 72.1798 11.1625Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M76.4932 9.39266V8.69349C76.4932 8.57083 76.4845 8.47446 76.4653 8.39911C76.4147 8.17481 76.303 7.98556 76.1302 7.83311C75.9574 7.68066 75.7498 7.60356 75.5107 7.60356C75.1634 7.60356 74.8911 7.74374 74.6922 8.01886C74.495 8.29572 74.3955 8.64969 74.3955 9.08075C74.3955 9.49605 74.4897 9.83249 74.68 10.0901C74.8807 10.3669 75.1512 10.5054 75.4932 10.5054C75.7969 10.5054 76.0412 10.3897 76.2262 10.1602C76.4042 9.94639 76.4932 9.69056 76.4932 9.39266ZM77.4555 11.1628H76.6318L76.5881 10.4882H76.562C76.3002 10.9998 75.8517 11.2557 75.2217 11.2557C74.7156 11.2557 74.2985 11.0577 73.9669 10.6616C73.6336 10.2639 73.4678 9.74869 73.4678 9.11611C73.4678 8.43621 73.6475 7.88774 74.0088 7.46718C74.3578 7.07642 74.7871 6.88016 75.2985 6.88016C75.8604 6.88016 76.2513 7.07116 76.4747 7.45141H76.4939V5.04199H77.4223V10.0308C77.4223 10.4408 77.4328 10.8158 77.4555 11.1628Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M84.4379 9.05993C84.4379 8.67267 84.3541 8.34149 84.1866 8.06637C83.9929 7.72993 83.7154 7.56346 83.3541 7.56346C82.9807 7.56346 82.6979 7.72993 82.5025 8.06637C82.3367 8.34149 82.2529 8.67968 82.2529 9.07745C82.2529 9.46471 82.3367 9.7959 82.5025 10.0728C82.7032 10.4092 82.9824 10.5757 83.3454 10.5757C83.6997 10.5757 83.9772 10.4057 84.1796 10.064C84.3506 9.78188 84.4379 9.44719 84.4379 9.05993ZM85.4001 9.02992C85.4001 9.67477 85.2168 10.204 84.8486 10.6193C84.4664 11.0451 83.9568 11.2571 83.3181 11.2571C82.7072 11.2571 82.2186 11.0538 81.8539 10.6438C81.4909 10.2355 81.3076 9.72033 81.3076 9.09826C81.3076 8.44815 81.4944 7.9172 81.8696 7.50191C82.2465 7.08836 82.7509 6.88159 83.3879 6.88159C84.0022 6.88159 84.4943 7.08486 84.866 7.49315C85.222 7.89092 85.4001 8.40259 85.4001 9.02992Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M90.3946 11.1625H89.4661V8.76006C89.4661 8.01884 89.1869 7.6491 88.6232 7.6491C88.3475 7.6491 88.1241 7.75074 87.9531 7.954C87.782 8.15727 87.6948 8.39734 87.6948 8.67245V11.1625H86.7663V8.16603C86.7663 7.79805 86.7559 7.39852 86.7314 6.9657H87.5482L87.5918 7.62107H87.6162C87.7262 7.4178 87.8868 7.25133 88.0997 7.11465C88.351 6.96045 88.6319 6.88159 88.9408 6.88159C89.3318 6.88159 89.6546 7.00776 89.9129 7.26009C90.234 7.57025 90.3946 8.03461 90.3946 8.64967V11.1625Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M96.6441 7.66664H95.6214V9.70282C95.6214 10.2215 95.8046 10.4791 96.1642 10.4791C96.3282 10.4791 96.4661 10.4668 96.5778 10.437L96.6022 11.145C96.4172 11.2133 96.1764 11.2484 95.8797 11.2484C95.5132 11.2484 95.227 11.1345 95.0193 10.9119C94.8134 10.6876 94.7122 10.3109 94.7122 9.78167V7.66664H94.0996V6.96747H94.7122V6.19821L95.6214 5.9231V6.96747H96.6441V7.66664Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M101.562 11.1623H100.632V8.77742C100.632 8.02393 100.352 7.64894 99.7905 7.64894C99.3594 7.64894 99.0645 7.86622 98.9057 8.30255C98.876 8.39542 98.8603 8.50581 98.8603 8.63899V11.1623H97.9336V5.0415H98.8603V7.57008H98.8777C99.1709 7.10923 99.5898 6.88143 100.134 6.88143C100.516 6.88143 100.836 7.00759 101.089 7.25992C101.403 7.57709 101.562 8.04671 101.562 8.66703V11.1623Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M105.748 8.62496C105.755 8.34809 105.694 8.10803 105.567 7.90826C105.408 7.64892 105.16 7.51925 104.828 7.51925C104.526 7.51925 104.277 7.64542 104.09 7.8995C103.935 8.10102 103.842 8.34284 103.814 8.62496H105.748ZM106.635 8.86693C106.635 9.03164 106.622 9.17358 106.598 9.28923H103.814C103.825 9.70278 103.959 10.0199 104.218 10.239C104.453 10.4335 104.759 10.5316 105.13 10.5316C105.542 10.5316 105.917 10.4668 106.256 10.3336L106.403 10.982C106.005 11.1537 105.54 11.2396 104.999 11.2396C104.352 11.2396 103.842 11.0485 103.474 10.6665C103.104 10.2828 102.921 9.76937 102.921 9.12452C102.921 8.49193 103.092 7.96624 103.437 7.54569C103.797 7.0971 104.282 6.8728 104.898 6.8728C105.497 6.8728 105.954 7.0971 106.263 7.54569C106.511 7.90316 106.635 8.34299 106.635 8.86693Z" fill="#FEFEFE"/>
-
-    </svg>
-  );
-};
