--- src/components/svgs/FaqIcon.tsx
@@ -1,37 +1,0 @@
-import React from 'react';
-
-interface FaqIconProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const FaqIcon: React.FC<FaqIconProps> = ({
-  className = '',
-  width = 24,
-  height = 24,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      fill="none"
-      viewBox="0 0 24 24"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path
-        fill="currentColor"
-        d="M12.01 2.975a8.99 8.99 0 0 1 6.346 2.646 8.997 8.997 0 0 1 2.641 6.381 9.093 9.093 0 0 1-1.252 4.575l1.305 4.372-4.63-1.102a8.982 8.982 0 0 1-4.417 1.175h-.066a8.99 8.99 0 0 1-6.347-2.645 8.998 8.998 0 0 1-2.64-6.382 8.998 8.998 0 0 1 2.645-6.38 9 9 0 0 1 6.35-2.64h.065Zm-.064 1.371c-1.982 0-3.992.836-5.383 2.24a7.723 7.723 0 0 0-2.245 5.409 7.732 7.732 0 0 0 2.24 5.412 7.727 7.727 0 0 0 5.383 2.244h.056a7.62 7.62 0 0 0 3.958-1.12l.241-.144 2.858.68-.801-2.688.167-.264a7.717 7.717 0 0 0 1.203-4.113 7.735 7.735 0 0 0-2.238-5.412A7.729 7.729 0 0 0 12 4.346h-.054Zm-.31 9.86c.686 0 1.12.518 1.12 1.12 0 .601-.434 1.119-1.12 1.119-.685 0-1.118-.518-1.118-1.12 0-.6.433-1.118 1.119-1.118Zm.202-6.646c1.746 0 2.903.843 2.903 2.265 0 .865-.409 1.576-1.444 2.3l-.482.337c-.325.229-.518.422-.518.76v.07h-1.373v-.096c0-.891.314-1.422 1.013-1.952l.253-.194c.566-.434.83-.784.83-1.24 0-.567-.408-.953-1.36-.953-.952 0-1.808.386-2.458.843 0-1.026.23-1.47.976-1.82.41-.193.964-.325 1.662-.325l-.002.005Z"
-      />
-    </svg>
-  );
-};
