--- src/components/svgs/GooglePlayDownIcon.tsx
@@ -1,67 +1,0 @@
-import React from 'react';
-
-interface GooglePlayDownIconProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const GooglePlayDownIcon: React.FC<GooglePlayDownIconProps> = ({
-  className = '',
-  width = 118,
-  height = 34,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      fill="none"
-      viewBox="0 0 118 34"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path fillRule="evenodd" clipRule="evenodd" d="M41.8827 9.04706C41.8827 9.78619 41.6618 10.3775 41.2235 10.8158C40.7244 11.3392 40.0757 11.6001 39.2757 11.6001C38.5122 11.6001 37.8601 11.3358 37.3262 10.8053C36.7905 10.2749 36.5244 9.61576 36.5244 8.83141C36.5244 8.04706 36.7905 7.38967 37.3262 6.85924C37.8601 6.3288 38.5122 6.06445 39.2757 6.06445C39.6566 6.06445 40.0201 6.13924 40.3644 6.28532C40.7088 6.43489 40.9853 6.63141 41.194 6.87837L40.7279 7.34445C40.3766 6.92532 39.8931 6.71663 39.2757 6.71663C38.7192 6.71663 38.2357 6.91141 37.8305 7.30445C37.4235 7.6975 37.2183 8.20706 37.2183 8.83141C37.2183 9.4575 37.4235 9.96706 37.8305 10.3601C38.2357 10.7531 38.7192 10.9479 39.2757 10.9479C39.8688 10.9479 40.3627 10.7514 40.7575 10.3566C41.0131 10.0992 41.1609 9.74271 41.2009 9.28358H39.2757V8.6488H41.8462C41.8688 8.78619 41.8827 8.91837 41.8827 9.04706Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M41.8831 9.04739H41.7944C41.7926 9.77086 41.5822 10.3326 41.1613 10.7535L41.1596 10.7552C40.6778 11.2596 40.057 11.5117 39.2761 11.5117C38.5317 11.5117 37.9091 11.2578 37.3891 10.743C36.8691 10.2265 36.6135 9.59695 36.6135 8.83173C36.6135 8.06826 36.8691 7.43869 37.3891 6.92217C37.9091 6.40739 38.5317 6.15347 39.2761 6.15347C39.6448 6.15347 39.9961 6.22478 40.3283 6.36739C40.6639 6.51173 40.9283 6.69956 41.1248 6.93608L41.1944 6.87869L41.13 6.81608L40.6657 7.28217L40.7283 7.34478L40.7944 7.28739C40.4291 6.84739 39.9126 6.62652 39.2761 6.62826C38.6987 6.62652 38.1891 6.83347 37.7683 7.24043C37.3439 7.64913 37.13 8.18652 37.1317 8.83173C37.13 9.47869 37.3439 10.0161 37.7683 10.4248C38.1891 10.8317 38.6987 11.0387 39.2761 11.0387C39.8883 11.0387 40.4083 10.8317 40.8204 10.4196C41.0935 10.1448 41.25 9.76565 41.29 9.2926L41.297 9.19521H39.3648V8.73608H41.8465V8.64913L41.7578 8.66304C41.7804 8.79695 41.7944 8.92391 41.7944 9.04739H41.9717C41.9717 8.91347 41.9578 8.77434 41.9317 8.63173L41.9196 8.56043H39.1874V9.3726H41.2013V9.28391L41.1144 9.27695C41.0744 9.72043 40.9317 10.0561 40.6935 10.2943C40.3161 10.6735 39.8483 10.8596 39.2761 10.8613C38.7387 10.8596 38.2848 10.6752 37.89 10.2978C37.5004 9.92043 37.3074 9.44043 37.3074 8.83173C37.3074 8.22478 37.5004 7.74478 37.89 7.36739C38.2848 6.99 38.7387 6.80391 39.2761 6.80391C39.8761 6.80565 40.3248 7.00391 40.6604 7.40217L40.7213 7.47521L41.257 6.9413L41.3144 6.88391L41.2622 6.8213C41.0448 6.56217 40.7544 6.35869 40.3978 6.20565C40.0431 6.0526 39.6691 5.97608 39.2761 5.97608C38.4917 5.97434 37.8135 6.25086 37.2657 6.79869C36.7126 7.34304 36.4344 8.02826 36.4344 8.83173C36.4344 9.63695 36.7126 10.3222 37.2657 10.8683C37.8135 11.4143 38.4917 11.6909 39.2761 11.6891C40.0952 11.6909 40.7717 11.4178 41.2883 10.877L41.2239 10.8161L41.2865 10.8787C41.7439 10.4213 41.9717 9.80391 41.9717 9.04739H41.8831Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M45.9524 6.83354H43.5385V8.51354H45.7142V9.15006H43.5385V10.8318H45.9524V11.4822H42.8603V6.18311H45.9524V6.83354Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M45.9524 6.83361V6.74492H43.4516V8.60231H45.6272V9.06144H43.4516V10.9206H45.8637V11.3936H42.9472V6.27187H45.8637V6.83361H45.9524V6.74492V6.83361H46.0411V6.09448H42.7716V11.571H46.0411V10.7432H43.6272V9.23883H45.8029V8.42666H43.6272V6.92231H46.0411V6.83361H45.9524Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M48.8248 11.482H48.1413V6.8333H46.6631V6.18286H50.3048V6.8333H48.8248V11.482Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M48.8244 11.4821V11.3934H48.2297V6.74467H46.7514V6.27163H50.2175V6.74467H48.7357V11.4821H48.8244V11.3934V11.4821H48.9131V6.92206H50.3914V6.09424H46.574V6.92206H48.0557V11.5708H48.9131V11.4821H48.8244Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M52.9387 11.483H53.6204V6.18384H52.9387V11.483Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M52.9395 11.4821H53.0282V6.27163H53.5308V11.3934H52.9395V11.4821H53.0282H52.9395V11.5708H53.7082V6.09424H52.8508V11.5708H52.9395V11.4821Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M56.6403 11.482H55.9585V6.8333H54.4785V6.18286H58.1203V6.8333H56.6403V11.482Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M56.6403 11.4821V11.3934H56.0472V6.74467H54.5672V6.27163H58.0316V6.74467H56.5516V11.4821H56.6403V11.3934V11.4821H56.729V6.92206H58.2072V6.09424H54.3898V6.92206H55.8716V11.5708H56.729V11.4821H56.6403Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M61.6308 10.3534C62.0238 10.7516 62.5038 10.9482 63.0708 10.9482C63.6378 10.9482 64.1195 10.7516 64.5091 10.3534C64.9038 9.95512 65.1004 9.44903 65.1004 8.83164C65.1004 8.21599 64.9038 7.70816 64.5091 7.31164C64.1195 6.91338 63.6378 6.71686 63.0708 6.71686C62.5038 6.71686 62.0238 6.91338 61.6308 7.31164C61.2395 7.70816 61.043 8.21599 61.043 8.83164C61.043 9.44903 61.2395 9.95512 61.6308 10.3534ZM65.013 10.7969C64.493 11.3325 63.8425 11.6003 63.0704 11.6003C62.2947 11.6003 61.6495 11.3325 61.1278 10.7969C60.6078 10.263 60.3486 9.60731 60.3486 8.83165C60.3486 8.05774 60.6078 7.40209 61.1278 6.86818C61.6495 6.33078 62.2947 6.0647 63.0704 6.0647C63.8408 6.0647 64.4878 6.33252 65.0078 6.87165C65.5312 7.40905 65.7939 8.06296 65.7939 8.83165C65.7939 9.60731 65.533 10.263 65.013 10.7969Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M61.6304 10.3535L61.5678 10.4161C61.9747 10.8282 62.4808 11.0387 63.0704 11.0387C63.6582 11.0387 64.166 10.8282 64.573 10.4161C64.9817 10.0022 65.1886 9.46824 65.1886 8.83172C65.1886 8.19693 64.9817 7.66302 64.573 7.24911C64.166 6.83693 63.6582 6.6265 63.0704 6.62824C62.4808 6.6265 61.9747 6.83693 61.5678 7.24911C61.1591 7.66302 60.9538 8.19693 60.9538 8.83172C60.9538 9.46824 61.1591 10.0022 61.5678 10.4161L61.6947 10.2908C61.3191 9.91172 61.1312 9.42998 61.1312 8.83172C61.1312 8.23519 61.3191 7.75345 61.6947 7.37432H61.693C62.0721 6.99172 62.5243 6.80563 63.0704 6.80389C63.6182 6.80563 64.0704 6.99172 64.4478 7.37432C64.8217 7.75519 65.0078 8.23519 65.0112 8.83172C65.0078 9.42998 64.8217 9.90998 64.4478 10.2908C64.0704 10.6735 63.6182 10.8613 63.0704 10.8613C62.5243 10.8595 62.0721 10.6735 61.693 10.2908L61.6304 10.3535ZM65.0126 10.797L64.95 10.7361C64.4439 11.2561 63.8247 11.5117 63.07 11.5117C62.3152 11.5117 61.6961 11.2561 61.1917 10.7361L61.1882 10.7326L61.1917 10.7361C60.6856 10.2161 60.4352 9.58826 60.4352 8.83173C60.4352 8.07695 60.6856 7.44913 61.1917 6.92913C61.6961 6.40913 62.3152 6.15347 63.07 6.15347C63.8195 6.15347 64.4369 6.41086 64.9465 6.9326C65.4526 7.45434 65.7047 8.08217 65.7047 8.83173C65.7047 9.58826 65.4543 10.2161 64.95 10.7361L65.0752 10.8596C65.6126 10.31 65.8821 9.62652 65.8804 8.83173C65.8821 8.04217 65.6108 7.36217 65.0717 6.80913C64.5361 6.25608 63.8595 5.97434 63.07 5.97608C62.2734 5.97434 61.5987 6.25434 61.0647 6.80565C60.5274 7.35695 60.2578 8.03869 60.2595 8.83173C60.2578 9.62652 60.5274 10.31 61.0647 10.8596L61.0682 10.8613L61.0647 10.8596C61.5987 11.4109 62.2734 11.6909 63.07 11.6891C63.863 11.6909 64.5413 11.4109 65.0752 10.8596L65.0126 10.797Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M66.749 11.482V6.18286H67.5803L70.1542 10.3046H70.1837L70.1542 9.28373V6.18286H70.8359V11.482H70.1229L67.4307 7.16025H67.3994L67.4307 8.18286V11.482H66.749Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M66.749 11.4821H66.8377V6.27163H67.5281L70.1038 10.3951H70.2725L70.2429 9.28206V6.27163H70.7472V11.3934H70.1751L67.4777 7.07163H67.3107L67.342 8.18293V11.3934H66.749V11.4821H66.8377H66.749V11.5708H67.5177V8.17946L67.4881 7.15859L67.3994 7.16033V7.24728H67.4307V7.16033L67.3559 7.20728L70.0759 11.5708H70.9229V6.09424H70.0655V9.28554L70.0951 10.3082L70.1838 10.3047V10.2177H70.1542V10.3047L70.229 10.2595L67.6272 6.09424H66.6603V11.5708H66.749V11.4821Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M94.4483 26.4969H96.0952V15.4551H94.4483V26.4969ZM109.292 19.4325L107.403 24.2204H107.346L105.386 19.4325H103.608L106.551 26.123L104.873 29.8447H106.593L111.125 19.4325H109.292ZM99.9457 25.243C99.403 25.243 98.6535 24.9734 98.6535 24.3056C98.6535 23.4534 99.5909 23.1264 100.4 23.1264C101.125 23.1264 101.466 23.283 101.906 23.4951C101.777 24.5178 100.897 25.243 99.9457 25.243ZM100.144 19.1925C98.9509 19.1925 97.7161 19.7178 97.2048 20.8812L98.6657 21.4917C98.9787 20.8812 99.5596 20.6812 100.172 20.6812C101.024 20.6812 101.89 21.1943 101.906 22.1038V22.2169C101.607 22.0464 100.968 21.7908 100.186 21.7908C98.6083 21.7908 97.0048 22.6569 97.0048 24.276C97.0048 25.7543 98.297 26.7056 99.7474 26.7056C100.853 26.7056 101.466 26.2082 101.85 25.6256H101.906V26.4778H103.497V22.2447C103.497 20.2847 102.033 19.1925 100.144 19.1925ZM89.9596 20.7769H87.6152V16.9925H89.9596C91.1909 16.9925 91.89 18.0117 91.89 18.8847C91.89 19.7404 91.1909 20.7769 89.9596 20.7769ZM89.9166 15.4551H85.9688V26.4969H87.6157V22.3125H89.9166C91.7444 22.3125 93.5374 20.9908 93.5374 18.8847C93.5374 16.7769 91.7444 15.4551 89.9166 15.4551ZM68.3824 25.2447C67.245 25.2447 66.292 24.2917 66.292 22.9821C66.292 21.6586 67.245 20.6917 68.3824 20.6917C69.5076 20.6917 70.3894 21.6586 70.3894 22.9821C70.3894 24.2917 69.5076 25.2447 68.3824 25.2447ZM70.2746 20.0517H70.219C69.8486 19.6099 69.139 19.2117 68.2416 19.2117C66.3633 19.2117 64.6416 20.8621 64.6416 22.9821C64.6416 25.0882 66.3633 26.7247 68.2416 26.7247C69.139 26.7247 69.8486 26.3265 70.219 25.8708H70.2746V26.4099C70.2746 27.8482 69.5077 28.6169 68.2694 28.6169C67.259 28.6169 66.6329 27.8899 66.3773 27.2795L64.9407 27.8778C65.3512 28.8725 66.4486 30.0969 68.2694 30.0969C70.2051 30.0969 71.8416 28.9578 71.8416 26.1838V19.4395H70.2746V20.0517ZM72.9809 26.4969H74.6313V15.4551H72.9809V26.4969ZM77.0639 22.8534C77.0187 21.403 78.1874 20.6638 79.0257 20.6638C79.6813 20.6638 80.2361 20.9891 80.4187 21.4604L77.0639 22.8534ZM82.1857 21.6012C81.8726 20.763 80.9178 19.2117 78.9683 19.2117C77.0343 19.2117 75.4257 20.7334 75.4257 22.9682C75.4257 25.0725 77.0187 26.7247 79.1543 26.7247C80.8761 26.7247 81.8726 25.6725 82.2848 25.0604L81.003 24.2047C80.5787 24.8325 79.9943 25.2447 79.1543 25.2447C78.3143 25.2447 77.7178 24.8604 77.3317 24.1073L82.3561 22.0291L82.1857 21.6012ZM42.1683 20.3647V21.9578H45.9822C45.8674 22.8534 45.57 23.5091 45.1143 23.9647C44.5596 24.5195 43.69 25.1299 42.1683 25.1299C39.8204 25.1299 37.9857 23.2378 37.9857 20.8917C37.9857 18.5421 39.8204 16.6499 42.1683 16.6499C43.4361 16.6499 44.3613 17.1491 45.043 17.7891L46.1683 16.6638C45.2135 15.7543 43.9491 15.0569 42.1683 15.0569C38.9526 15.0569 36.25 17.6743 36.25 20.8917C36.25 24.1073 38.9526 26.7247 42.1683 26.7247C43.9039 26.7247 45.2135 26.156 46.2378 25.0882C47.29 24.0343 47.6187 22.556 47.6187 21.3595C47.6187 20.9891 47.5891 20.6482 47.5335 20.3647H42.1683ZM51.9547 25.2447C50.8174 25.2447 49.833 24.3056 49.833 22.9682C49.833 21.6169 50.8174 20.6917 51.9547 20.6917C53.0921 20.6917 54.0747 21.6169 54.0747 22.9682C54.0747 24.3056 53.0921 25.2447 51.9547 25.2447ZM51.9548 19.2117C49.8765 19.2117 48.1826 20.7908 48.1826 22.9682C48.1826 25.1299 49.8765 26.7247 51.9548 26.7247C54.0313 26.7247 55.7252 25.1299 55.7252 22.9682C55.7252 20.7908 54.0313 19.2117 51.9548 19.2117ZM60.1804 25.2447C59.0413 25.2447 58.0587 24.3056 58.0587 22.9682C58.0587 21.6169 59.0413 20.6917 60.1804 20.6917C61.3178 20.6917 62.297 21.6169 62.297 22.9682C62.297 24.3056 61.3178 25.2447 60.1804 25.2447ZM60.1804 19.2117C58.1022 19.2117 56.41 20.7908 56.41 22.9682C56.41 25.1299 58.1022 26.7247 60.1804 26.7247C62.257 26.7247 63.9491 25.1299 63.9491 22.9682C63.9491 20.7908 62.257 19.2117 60.1804 19.2117Z" fill="#FEFEFE"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M9.2175 6.66615C8.96011 6.93746 8.81055 7.35833 8.81055 7.90615V27.4227C8.81055 27.9705 8.96011 28.3914 9.2175 28.6627L9.28185 28.727L20.2158 17.7931V17.5357L9.28185 6.60181L9.2175 6.66615Z" fill="#5AC8F2"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M20.2158 17.5351V17.7942L23.8593 21.4395L23.941 21.3908L28.261 18.9386C29.4941 18.236 29.4941 17.0916 28.261 16.389L23.861 13.8899L20.2158 17.5351Z" fill="url(#paint0_linear_196_3101)"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M9.21786 28.6621C9.62308 29.0934 10.2961 29.1456 11.0526 28.7177L23.9413 21.3908L20.2161 17.6638L9.21786 28.6621Z" fill="url(#paint1_linear_196_3101)"/>
-      <path fillRule="evenodd" clipRule="evenodd" d="M9.21786 6.66722L20.2161 17.6637L23.9413 13.9368L11.0526 6.61157C10.6996 6.41157 10.3622 6.31592 10.0631 6.31592C9.72395 6.31592 9.43352 6.43766 9.21786 6.66722Z" fill="url(#paint2_linear_196_3101)"/>
-      <defs>
-      <linearGradient id="paint0_linear_196_3101" x1="29.1858" y1="13.8899" x2="20.2158" y2="13.8899" gradientUnits="userSpaceOnUse">
-      <stop stopColor="#FAD961"/>
-      <stop offset="1" stopColor="#F99B1B"/>
-      </linearGradient>
-      <linearGradient id="paint1_linear_196_3101" x1="8.0793" y1="11.1382" x2="-5.98117" y2="34.8009" gradientUnits="userSpaceOnUse">
-      <stop stopColor="#EE4547"/>
-      <stop offset="1" stopColor="#C71A6D"/>
-      </linearGradient>
-      <linearGradient id="paint2_linear_196_3101" x1="-9.70506" y1="2.59557" x2="1.77214" y2="21.9166" gradientUnits="userSpaceOnUse">
-      <stop stopColor="#73C169"/>
-      <stop offset="1" stopColor="#289E6F"/>
-      </linearGradient>
-      </defs>
-
-    </svg>
-  );
-};
