--- src/components/svgs/InstagramIcon.tsx
@@ -1,45 +1,0 @@
-import React from 'react';
-
-interface InstagramIconProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const InstagramIcon: React.FC<InstagramIconProps> = ({
-  className = '',
-  width = 32,
-  height = 32,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      fill="none"
-      viewBox="0 0 32 32"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path
-        fill="currentColor"
-        d="M16 7.98948C18.5917 7.98948 18.9451 7.98948 20.0054 7.98948C20.9478 7.98948 21.419 8.22509 21.7724 8.34289C22.2436 8.5785 22.5971 8.69631 22.9505 9.04972C23.3039 9.40314 23.5395 9.75655 23.6573 10.2278C23.7751 10.5812 23.8929 11.0524 24.0107 11.9948C24.0107 13.0551 24.0107 13.2907 24.0107 16.0002C24.0107 18.7097 24.0107 18.9453 24.0107 20.0055C24.0107 20.948 23.7751 21.4192 23.6573 21.7726C23.4217 22.2438 23.3039 22.5972 22.9505 22.9507C22.5971 23.3041 22.2436 23.5397 21.7724 23.6575C21.419 23.7753 20.9478 23.8931 20.0054 24.0109C18.9451 24.0109 18.7095 24.0109 16 24.0109C13.2905 24.0109 13.0549 24.0109 11.9947 24.0109C11.0522 24.0109 10.581 23.7753 10.2276 23.6575C9.75637 23.4219 9.40295 23.3041 9.04954 22.9507C8.69613 22.5972 8.46052 22.2438 8.34271 21.7726C8.22491 21.4192 8.1071 20.948 7.9893 20.0055C7.9893 18.9453 7.9893 18.7097 7.9893 16.0002C7.9893 13.2907 7.9893 13.0551 7.9893 11.9948C7.9893 11.0524 8.22491 10.5812 8.34271 10.2278C8.57832 9.75655 8.69613 9.40314 9.04954 9.04972C9.40295 8.69631 9.75637 8.4607 10.2276 8.34289C10.581 8.22509 11.0522 8.10729 11.9947 7.98948C13.0549 7.98948 13.4083 7.98948 16 7.98948ZM16 6.22241C13.2905 6.22241 13.0549 6.22241 11.9947 6.22241C10.9344 6.22241 10.2276 6.45802 9.63856 6.69363C9.04954 6.92924 8.46052 7.28265 7.87149 7.87168C7.28247 8.4607 7.04686 8.93192 6.69345 9.63875C6.45784 10.2278 6.34003 10.9346 6.22223 11.9948C6.22223 13.0551 6.22223 13.4085 6.22223 16.0002C6.22223 18.7097 6.22223 18.9453 6.22223 20.0055C6.22223 21.0658 6.45784 21.7726 6.69345 22.3616C6.92906 22.9507 7.28247 23.5397 7.87149 24.1287C8.46052 24.7177 8.93173 24.9533 9.63856 25.3068C10.2276 25.5424 10.9344 25.6602 11.9947 25.778C13.0549 25.778 13.4083 25.778 16 25.778C18.5917 25.778 18.9451 25.778 20.0054 25.778C21.0656 25.778 21.7724 25.5424 22.3615 25.3068C22.9505 25.0711 23.5395 24.7177 24.1285 24.1287C24.7175 23.5397 24.9532 23.0685 25.3066 22.3616C25.5422 21.7726 25.66 21.0658 25.7778 20.0055C25.7778 18.9453 25.7778 18.5919 25.7778 16.0002C25.7778 13.4085 25.7778 13.0551 25.7778 11.9948C25.7778 10.9346 25.5422 10.2278 25.3066 9.63875C25.071 9.04972 24.7175 8.4607 24.1285 7.87168C23.5395 7.28265 23.0683 7.04704 22.3615 6.69363C21.7724 6.45802 21.0656 6.34022 20.0054 6.22241C18.9451 6.22241 18.7095 6.22241 16 6.22241Z"
-      />
-      <path
-        fill="currentColor"
-        d="M16 10.9346C13.1727 10.9346 10.9344 13.1729 10.9344 16.0002C10.9344 18.8275 13.1727 21.0658 16 21.0658C18.8273 21.0658 21.0656 18.8275 21.0656 16.0002C21.0656 13.1729 18.8273 10.9346 16 10.9346ZM16 19.2987C14.2329 19.2987 12.7015 17.8851 12.7015 16.0002C12.7015 14.2331 14.1151 12.7017 16 12.7017C17.7671 12.7017 19.2985 14.1153 19.2985 16.0002C19.2985 17.7673 17.7671 19.2987 16 19.2987Z"
-      />
-      <path
-        fill="currentColor"
-        d="M21.1834 11.9948C21.834 11.9948 22.3615 11.4674 22.3615 10.8168C22.3615 10.1662 21.834 9.63875 21.1834 9.63875C20.5328 9.63875 20.0054 10.1662 20.0054 10.8168C20.0054 11.4674 20.5328 11.9948 21.1834 11.9948Z"
-      />
-    </svg>
-  );
-};
