--- src/components/svgs/KoreanAirLogo.tsx
@@ -1,128 +1,0 @@
-import React from 'react';
-
-interface KoreanAirLogoProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const KoreanAirLogo: React.FC<KoreanAirLogoProps> = ({
-  className = '',
-  width = 205,
-  height = 40,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      width={width}
-      height={height}
-      viewBox="0 0 205 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <g id="Logo">
-        <g id="logo">
-          <g id="Group 24">
-            <mask
-              id="mask0_407_41521"
-              style={{ maskType: 'luminance' }}
-              maskUnits="userSpaceOnUse"
-              x="5"
-              y="5"
-              width="29"
-              height="30"
-            >
-              <path
-                id="Clip 23"
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M5.12823 5.12817H33.3101V34.8717H5.12823V5.12817Z"
-                fill="white"
-              />
-            </mask>
-            <g mask="url(#mask0_407_41521)">
-              <path
-                id="Fill 22"
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M7.09128 27.302L6.97224 27.3839C9.39639 31.8528 13.9844 34.8718 19.2191 34.8718C27.0045 34.8718 33.3101 28.2092 33.3101 20.0001C33.3101 17.7394 32.3732 15.8359 31.2653 14.3933C29.793 12.4748 27.6961 11.5528 25.6066 11.5379C23.8815 11.5231 22.097 12.1031 20.8401 13.2407C19.405 14.542 19.0853 16.2152 18.7135 18.9292C18.3046 21.9036 15.9102 23.9559 12.8318 23.9559C8.84609 23.9559 6.37747 21.5021 6.54097 17.7469C6.55592 17.4866 6.57815 17.2265 6.62282 16.9737C7.37381 12.5419 12.1403 7.96884 19.3084 7.9836C24.1194 7.9836 27.7332 9.06178 31.3471 12.6979L31.4662 12.6163C29.042 8.14729 24.4615 5.12817 19.2191 5.12817C11.4412 5.12817 5.12811 11.7907 5.12811 20.0001C5.12811 22.2605 6.06503 24.1641 7.17312 25.6066C8.63792 27.5251 10.7424 28.4396 12.8318 28.4621C14.5569 28.4843 16.3415 27.8968 17.5981 26.7592C19.0334 25.4579 19.3531 23.7849 19.7248 21.0708C20.1338 18.0963 22.5355 16.044 25.6066 16.044C29.5996 16.044 32.0684 18.4979 31.9047 22.253C31.89 22.5133 31.8675 22.7736 31.8231 23.0265C31.0571 27.4582 26.3131 32.0313 19.13 32.0163C14.3189 32.0163 10.7052 30.9382 7.09128 27.302"
-                fill="currentColor"
-              />
-            </g>
-          </g>
-          <path
-            id="Fill 25"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M59.821 27.3335C60.2617 27.7982 60.6632 28.1643 61.0257 28.4319H58.7346C57.9316 28.4319 57.2309 28.29 56.6325 28.0067C56.034 27.7232 55.4517 27.2354 54.8847 26.5423L49.1927 19.4332L56.5382 11.5682H60.3171L52.948 19.2915L58.0733 25.3378C58.7975 26.204 59.3802 26.8693 59.821 27.3335V27.3335ZM45.3428 28.4319H48.6966V11.5682H45.3428V28.4319Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 27"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M72.7879 25.3614C73.6223 24.8891 74.2759 24.1964 74.7482 23.2829C75.2207 22.3699 75.4568 21.2754 75.4568 20C75.4568 18.7246 75.2207 17.6303 74.7482 16.7171C74.2759 15.804 73.6223 15.1109 72.7879 14.6386C71.9531 14.1663 71.0005 13.9301 69.93 13.9301C68.8435 13.9301 67.883 14.1663 67.0485 14.6386C66.2137 15.1109 65.5605 15.804 65.0882 16.7171C64.6157 17.6303 64.3796 18.7246 64.3796 20C64.3796 21.2754 64.6157 22.3699 65.0882 23.2829C65.5605 24.1964 66.2137 24.8891 67.0485 25.3614C67.883 25.8337 68.8435 26.0699 69.93 26.0699C71.0005 26.0699 71.9531 25.8337 72.7879 25.3614M65.3243 27.6169C63.9546 26.8849 62.8799 25.8575 62.1003 24.5347C61.3211 23.2121 60.9313 21.7006 60.9313 20C60.9313 18.2994 61.3211 16.7879 62.1003 15.4653C62.8799 14.1425 63.9546 13.1151 65.3243 12.3829C66.6943 11.6509 68.2294 11.2848 69.93 11.2848C71.6146 11.2848 73.1421 11.6509 74.512 12.3829C75.8819 13.1151 76.9565 14.1467 77.736 15.477C78.5154 16.8078 78.9051 18.3153 78.9051 20C78.9051 21.685 78.5154 23.1926 77.736 24.5228C76.9565 25.8536 75.8819 26.8849 74.512 27.6169C73.1421 28.3491 71.6146 28.7152 69.93 28.7152C68.2294 28.7152 66.6943 28.3491 65.3243 27.6169"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 29"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M97.5401 28.4319H95.4144C94.5482 28.4319 93.816 28.2823 93.218 27.9831C92.6194 27.6841 92.0683 27.2038 91.5646 26.5423L86.4631 19.6929H87.9982C89.2577 19.6929 90.2102 19.4531 90.8561 18.9726C91.5015 18.4926 91.8245 17.788 91.8245 16.8587C91.8245 15.9458 91.5055 15.2686 90.8678 14.8275C90.2301 14.3869 89.2499 14.1662 87.9273 14.1662H85.8254V28.4319H82.4714V11.5682H88.6122C89.9504 11.5682 91.1195 11.7729 92.1197 12.1824C93.1195 12.592 93.8869 13.1783 94.4225 13.942C94.9575 14.7058 95.2254 15.5992 95.2254 16.6226C95.2254 17.7564 94.8984 18.7447 94.2454 19.5867C93.5917 20.4291 92.6746 21.0551 91.4938 21.4644L94.5877 25.3378C95.6425 26.6604 96.6268 27.692 97.5401 28.4319"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 31"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M110.719 25.7865V28.4319H99.6185V11.5682H110.483V14.2134H102.972V18.4176H109.279V21.0156H102.972V25.7865H110.719Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 33"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M123.756 22.4564L121.064 16.2447L118.371 22.4564H123.756ZM130.275 28.4317H128.268C127.606 28.4317 127.055 28.2784 126.614 27.9713C126.173 27.6643 125.779 27.1174 125.433 26.3298L124.866 25.0072H117.261L115.773 28.4317H112.608L120.355 11.5681H122.197L128.22 24.6528C129.071 26.495 129.756 27.7551 130.275 28.4317V28.4317Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 35"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M147.564 11.5682V28.4319H145.722L135.424 17.3312V28.4319H132.542V11.5682H134.385L144.682 22.6689V11.5682H147.564Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 37"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M167.734 22.4564L165.041 16.2447L162.349 22.4564H167.734ZM174.253 28.4317H172.245C171.584 28.4317 171.032 28.2784 170.591 27.9713C170.151 27.6643 169.757 27.1174 169.411 26.3298L168.844 25.0072H161.239L159.751 28.4317H156.586L164.333 11.5681H166.175L172.198 24.6528C173.048 26.495 173.733 27.7551 174.253 28.4317V28.4317Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 39"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M176.52 28.4317H179.874V11.5681H176.52V28.4317Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 41"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M199.996 28.4319H197.871C197.005 28.4319 196.272 28.2823 195.674 27.9831C195.075 27.6841 194.524 27.2038 194.021 26.5423L188.919 19.6929H190.454C191.714 19.6929 192.666 19.4531 193.312 18.9726C193.958 18.4926 194.281 17.788 194.281 16.8587C194.281 15.9458 193.962 15.2686 193.324 14.8275C192.686 14.3869 191.706 14.1662 190.383 14.1662H188.281V28.4319H184.928V11.5682H191.068C192.407 11.5682 193.576 11.7729 194.576 12.1824C195.576 12.592 196.343 13.1783 196.879 13.942C197.414 14.7058 197.682 15.5992 197.682 16.6226C197.682 17.7564 197.355 18.7447 196.701 19.5867C196.048 20.4291 195.131 21.0551 193.95 21.4644L197.044 25.3378C198.099 26.6604 199.083 27.692 199.996 28.4319"
-            fill="currentColor"
-          />
-        </g>
-      </g>
-    </svg>
-  );
-};
