--- src/components/svgs/MobileKoreanAirLogo.tsx
@@ -1,70 +1,0 @@
-import React from 'react';
-
-interface MobileKoreanAirLogoProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const MobileKoreanAirLogo: React.FC<MobileKoreanAirLogoProps> = ({
-  className = '',
-  width = 145,
-  height = 22,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      xmlns="http://www.w3.org/2000/svg"
-      width={width}
-      height={height}
-      fill="none"
-      viewBox="0 0 145 22"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path
-        d="M1.45124 16.4004L1.36337 16.4614C3.15615 19.7663 6.54757 22.0007 10.418 22.0007C16.1748 22.0007 20.836 17.0724 20.836 11C20.836 9.32841 20.143 7.92054 19.3238 6.85293C18.2354 5.43372 16.685 4.75176 15.1402 4.74041C13.8647 4.72907 12.5452 5.15866 11.6156 5.99942C10.5541 6.9621 10.3174 8.19984 10.0439 10.2074C9.74198 12.4079 7.97046 13.9249 5.6944 13.9249C2.74799 13.9249 0.922611 12.1101 1.04308 9.33267C1.05441 9.13985 1.07 8.94702 1.10402 8.76129C1.65957 5.48334 5.18278 2.10047 10.4832 2.11181C14.0404 2.11181 16.7119 2.91003 19.3848 5.5996L19.4726 5.53863C17.6813 2.23233 14.2941 -0.000709534 10.418 -0.000709534C4.66833 -0.000709534 0 4.92756 0 11C0 12.6716 0.693021 14.0795 1.51218 15.1471C2.59493 16.5663 4.15104 17.2426 5.69581 17.2596C6.97131 17.2766 8.29075 16.8413 9.22045 16.0006C10.2819 15.0379 10.5186 13.8002 10.7921 11.7925C11.094 9.59212 12.8698 8.07508 15.1416 8.07508C18.0937 8.07508 19.9191 9.88986 19.7986 12.6673C19.7873 12.8602 19.7717 13.053 19.7376 13.2387C19.1708 16.5167 15.6631 19.8995 10.3528 19.8882C6.79558 19.8882 4.12412 19.09 1.45124 16.4004Z"
-        fill="currentColor"
-      />
-      <path
-        d="M29.7334 4.7631H32.2135V17.2369H29.7334V4.7631ZM40.4377 16.4245C40.7636 16.7676 41.0598 17.0384 41.3277 17.2369H39.6341C39.0403 17.2369 38.5216 17.132 38.0794 16.9222C37.6372 16.7123 37.2064 16.3522 36.7869 15.839L32.5778 10.5803L38.0086 4.7631H40.8019L35.3541 10.4754L39.1438 14.9472C39.6795 15.588 40.1103 16.08 40.4363 16.4231L40.4377 16.4245Z"
-        fill="currentColor"
-      />
-      <path
-        d="M44.5078 16.6343C43.4945 16.0927 42.7008 15.3328 42.124 14.3545C41.5472 13.3762 41.2595 12.2576 41.2595 11C41.2595 9.74241 41.5472 8.62377 42.124 7.64549C42.7008 6.66721 43.4945 5.90727 44.5078 5.36567C45.5211 4.82407 46.6563 4.55327 47.9134 4.55327C49.1705 4.55327 50.2886 4.82407 51.3005 5.36567C52.3139 5.90727 53.1075 6.67004 53.6843 7.654C54.2611 8.63795 54.5488 9.75376 54.5488 11C54.5488 12.2462 54.2611 13.3621 53.6843 14.346C53.1075 15.33 52.3139 16.0927 51.3005 16.6343C50.2872 17.1759 49.1577 17.4467 47.9134 17.4467C46.6691 17.4467 45.5211 17.1759 44.5078 16.6343ZM50.0265 14.9656C50.6429 14.6168 51.1262 14.1036 51.4763 13.4287C51.8249 12.7538 52.0006 11.9443 52.0006 11C52.0006 10.0557 51.8263 9.2476 51.4763 8.57131C51.1276 7.89644 50.6444 7.3832 50.0265 7.03442C49.4085 6.68564 48.7056 6.50983 47.9134 6.50983C47.1211 6.50983 46.3998 6.68422 45.7833 7.03442C45.1654 7.3832 44.6835 7.89644 44.3335 8.57131C43.9848 9.24619 43.8091 10.0557 43.8091 11C43.8091 11.9443 43.9834 12.7524 44.3335 13.4287C44.6821 14.105 45.1654 14.6168 45.7833 14.9656C46.3998 15.3144 47.1098 15.4902 47.9134 15.4902C48.7169 15.4902 49.4085 15.3158 50.0265 14.9656Z"
-        fill="currentColor"
-      />
-      <path
-        d="M68.3272 17.2369H66.7555C66.1149 17.2369 65.5736 17.1263 65.1314 16.9051C64.6892 16.684 64.281 16.3281 63.9083 15.839L60.1357 10.7732H61.2709C62.202 10.7732 62.9063 10.5959 63.3839 10.2401C63.8615 9.88561 64.0996 9.36386 64.0996 8.67623C64.0996 7.98859 63.8644 7.49945 63.3924 7.17336C62.9205 6.84727 62.1963 6.68422 61.2184 6.68422H59.6637V17.2369H57.1836V4.7631H61.7244C62.7136 4.7631 63.5781 4.9148 64.3179 5.2168C65.0577 5.5202 65.6246 5.95405 66.02 6.51834C66.4154 7.08404 66.6138 7.74473 66.6138 8.50184C66.6138 9.33976 66.3714 10.0713 65.8896 10.6938C65.4063 11.3176 64.7289 11.7798 63.8545 12.0832L66.1419 14.9486C66.9213 15.9269 67.6498 16.6896 68.3244 17.2369H68.3272Z"
-        fill="currentColor"
-      />
-      <path
-        d="M78.0719 15.2803V17.2369H69.8647V4.7631H77.8976V6.71966H72.3449V9.8289H77.0075V11.75H72.3449V15.2789H78.0733L78.0719 15.2803Z"
-        fill="currentColor"
-      />
-      <path
-        d="M92.5303 17.2369H91.0465C90.5575 17.2369 90.1494 17.1235 89.8234 16.8966C89.4975 16.6698 89.2055 16.2643 88.9504 15.683L88.5309 14.7047H82.9074L81.8076 17.2383H79.4678L85.1962 4.76451H86.5581L91.0111 14.4424C91.6403 15.8049 92.1463 16.7364 92.5303 17.2383V17.2369ZM87.7103 12.8176L85.7191 8.22253L83.7279 12.8176H87.7089H87.7103Z"
-        fill="currentColor"
-      />
-      <path
-        d="M105.314 4.7631V17.2369H103.952L96.3386 9.02642V17.2369H94.2085V4.7631H95.5704L103.184 12.9736V4.7631H105.314Z"
-        fill="currentColor"
-      />
-      <path
-        d="M125.047 17.2369H123.563C123.074 17.2369 122.666 17.1235 122.34 16.8966C122.014 16.6698 121.722 16.2643 121.467 15.683L121.048 14.7047H115.424L114.324 17.2383H111.985L117.713 4.76451H119.075L123.528 14.4424C124.157 15.8049 124.663 16.7364 125.047 17.2383V17.2369ZM120.227 12.8176L118.236 8.22253L116.245 12.8176H120.226H120.227Z"
-        fill="currentColor"
-      />
-      <path d="M126.724 4.7631H129.204V17.2369H126.724V4.7631Z" fill="currentColor" />
-      <path
-        d="M144.08 17.2369H142.509C141.868 17.2369 141.327 17.1263 140.885 16.9051C140.442 16.684 140.034 16.3281 139.661 15.839L135.889 10.7732H137.024C137.955 10.7732 138.66 10.5959 139.137 10.2401C139.615 9.88561 139.853 9.36386 139.853 8.67623C139.853 7.98859 139.618 7.49945 139.146 7.17336C138.674 6.84727 137.949 6.68422 136.972 6.68422H135.417V17.2369H132.937V4.7631H137.478C138.467 4.7631 139.331 4.9148 140.071 5.2168C140.811 5.5202 141.378 5.95405 141.773 6.51834C142.169 7.08404 142.367 7.74473 142.367 8.50184C142.367 9.33976 142.125 10.0713 141.643 10.6938C141.159 11.3176 140.482 11.7798 139.608 12.0832L141.895 14.9486C142.674 15.9269 143.403 16.6896 144.078 17.2369H144.08Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
