--- src/components/svgs/SignUpIcon.tsx
@@ -1,37 +1,0 @@
-import React from 'react';
-
-interface SignUpIconProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const SignUpIcon: React.FC<SignUpIconProps> = ({
-  className = '',
-  width = 24,
-  height = 24,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      width={width}
-      height={height}
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <path
-        d="M17.1318 12.2295C18.4112 12.2244 19.6857 12.7524 20.5869 13.6611C21.4956 14.5597 22.0235 15.8369 22.0186 17.1162C22.0236 18.3956 21.4956 19.6701 20.5869 20.5713C19.6883 21.4801 18.4113 22.008 17.1318 22.0029C15.8524 22.008 14.578 21.4801 13.6768 20.5713C12.768 19.6726 12.24 18.3956 12.2451 17.1162C12.2401 15.8369 12.768 14.5622 13.6768 13.6611C14.5754 12.7525 15.8525 12.2244 17.1318 12.2295ZM17.1318 13.623C15.2025 13.623 13.6357 15.1898 13.6357 17.1191C13.6357 19.0485 15.2 20.6152 17.1318 20.6152C19.0611 20.6152 20.6269 19.0484 20.627 17.1191C20.627 15.1898 19.0611 13.6231 17.1318 13.623ZM10.166 3.2002C11.9748 3.20023 13.4812 4.55199 13.666 6.34277C13.8097 7.71792 13.1602 9.01618 11.9775 9.72949L5.7627 13.5527C5.18563 13.8939 4.77786 14.4199 4.61621 15.0381C4.59569 15.1074 4.58789 15.1823 4.58789 15.2695V17.832H11.3203C11.3203 17.832 11.3236 18.2002 11.3799 18.5254C11.4433 18.8915 11.5948 19.2078 11.5996 19.2178H3.2002V15.2285C3.2002 15.0977 3.21764 14.8409 3.25098 14.71C3.51524 13.7171 4.16713 12.8497 5.04199 12.334L11.3438 8.52637C12.1338 8.01326 12.495 7.08478 12.2617 6.16895C11.9846 5.08368 10.9204 4.39568 9.78125 4.5752C8.70107 4.74453 7.97754 5.6146 7.97754 6.74609C7.97771 7.75934 8.92409 8.67492 9.4834 9.05469L9.79688 9.26758L9.41895 9.28613C8.6726 9.31682 7.97237 9.08328 7.54395 8.66016C6.76156 7.88546 6.69729 7.11304 6.69727 6.45117C6.69727 4.56279 8.49821 3.22073 10.1172 3.2002H10.166ZM17.8164 16.4316H19.0146V17.8018H17.8164V19H16.4463V17.8018H15.248V16.4316H16.4463V15.2334H17.8164V16.4316ZM11.1699 11.166C11.7887 11.2192 13.0848 11.3931 14.3945 11.9053C14.0754 12.1068 13.773 12.3341 13.4902 12.5889H13.4873C13.283 12.4685 13.0704 12.3446 12.8438 12.2158C12.757 12.1682 11.7909 11.6148 11.4102 11.416C11.371 11.3936 10.8623 11.1392 10.8613 11.1387L11.1699 11.166Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-};
