--- src/components/svgs/SkyTeamLogo.tsx
@@ -1,128 +1,0 @@
-import React from 'react';
-
-interface SkyTeamLogoProps {
-  className?: string;
-  width?: number;
-  height?: number;
-  role?: string;
-  'aria-hidden'?: boolean;
-}
-
-export const SkyTeamLogo: React.FC<SkyTeamLogoProps> = ({
-  className = '',
-  width = 38,
-  height = 40,
-  role = 'none',
-  'aria-hidden': ariaHidden = true,
-  ...props
-}) => {
-  return (
-    <svg
-      className={className}
-      width={width}
-      height={height}
-      viewBox="0 0 38 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      role={role}
-      aria-hidden={ariaHidden}
-      {...props}
-    >
-      <g id="skyteam">
-        <g id="skyteam_2">
-          <path
-            id="Fill 1"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M26.9619 8.39248L26.9891 8.31389L27.0595 8.34092C27.084 8.35184 27.1247 8.36813 27.111 8.40609C27.0974 8.44404 27.0731 8.44404 27.0405 8.42775L26.9619 8.39248ZM27.0568 8.48467C27.0948 8.50365 27.1409 8.48736 27.1599 8.44941L27.1626 8.44135C27.1762 8.41145 27.1762 8.37618 27.1572 8.34916C27.1383 8.33018 27.1137 8.31389 27.0867 8.30565L26.9483 8.2514L26.8398 8.53355L26.8941 8.55521L26.9376 8.43867L26.9945 8.46033L27.0108 8.60121L27.0758 8.62575L27.0568 8.48467ZM26.9213 8.65009C26.8019 8.6039 26.7449 8.47106 26.7911 8.35184C26.8344 8.23511 26.9646 8.17818 27.0813 8.22169L27.0867 8.22438C27.2032 8.27038 27.2629 8.4034 27.2142 8.52263C27.1707 8.63648 27.0405 8.69629 26.9213 8.65009V8.65009ZM26.9023 8.69629C27.0432 8.75321 27.2059 8.68536 27.2629 8.54429L27.2655 8.53892C27.3227 8.39516 27.2548 8.23243 27.111 8.17281C26.9673 8.11589 26.8047 8.18374 26.7449 8.32731C26.6853 8.47106 26.7559 8.6338 26.9023 8.69629C26.8996 8.69341 26.8996 8.69341 26.9023 8.69629V8.69629Z"
-            fill="currentColor"
-          />
-          <g id="Group 5">
-            <mask
-              id="mask0_407_41597"
-              style={{ maskType: 'luminance' }}
-              maskUnits="userSpaceOnUse"
-              x="12"
-              y="12"
-              width="21"
-              height="22"
-            >
-              <path
-                id="Clip 4"
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M12.3845 12.3654H32.5875V33.5033H12.3845V12.3654Z"
-                fill="white"
-              />
-            </mask>
-            <g mask="url(#mask0_407_41597)">
-              <path
-                id="Fill 3"
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M28.3853 12.3654L28.3744 12.3763C33.0521 17.0843 33.0272 24.6908 28.3219 29.3683C24.007 33.6554 17.168 34.0428 12.3981 30.2675L12.3845 30.2784C17.2383 34.802 24.8395 34.534 29.3631 29.6802C33.8759 24.8373 33.6216 17.2551 28.7894 12.7261C28.6592 12.604 28.5237 12.4819 28.3853 12.3654"
-                fill="currentColor"
-              />
-            </g>
-          </g>
-          <path
-            id="Fill 6"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M7.08872 26.0206C6.70901 26.2349 6.61681 26.5548 6.8175 27.0511C6.99633 27.4984 7.28654 27.6991 7.59035 27.5772C7.83435 27.477 7.87499 27.268 7.77206 26.7285L7.72874 26.5195C7.63654 26.197 7.61756 25.8579 7.67449 25.527C7.72337 25.2913 7.88879 25.0986 8.11382 25.0118C8.65627 24.795 9.16881 25.1259 9.47512 25.8906C9.80864 26.7124 9.62961 27.3441 8.98423 27.6098L8.82149 27.2058C9.26619 27.0106 9.38023 26.5903 9.14696 25.9936C8.94896 25.5 8.6371 25.2859 8.31182 25.4161C8.04596 25.5217 7.98099 25.7686 8.08123 26.3217L8.11918 26.5548L8.15177 26.7584C8.21407 26.9997 8.22767 27.2491 8.18953 27.496C8.14084 27.7156 7.98079 27.8918 7.76918 27.9677C7.26219 28.1657 6.76594 27.8701 6.48935 27.1839C6.1911 26.4327 6.34846 25.8822 6.92331 25.6275L7.08872 26.0206Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 8"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M6.74462 21.6871L8.56978 20.1497L8.58339 20.7191L7.03214 21.9963L7.53913 22.5468L8.62383 22.5224L8.63457 22.97L5.53245 23.0348L5.51884 22.5928L6.73906 22.5658H7.06722L6.95873 22.4573L6.85848 22.3461L5.48645 20.8627L5.47015 20.2742L6.74462 21.6871Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 10"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M8.12524 16.89L9.31575 17.2996L9.16931 17.7253L7.9788 17.2994L5.83374 17.9501L6.01008 17.4349L7.34416 17.0717L7.50957 17.0228C7.5828 17.0012 7.65889 16.9849 7.7348 16.9712C7.65065 16.8737 7.57206 16.7732 7.49884 16.6676L6.67175 15.5503L6.84273 15.0702L8.12524 16.89Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 12"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M10.2191 11.1981L9.56547 11.998L11.5233 13.5955L11.0569 14.1757L9.1018 12.5757L8.44817 13.3647L8.00616 13.0095L9.78532 10.8374L10.2191 11.1981Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 14"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M14.3954 8.26151L12.8308 9.17256L13.167 9.74471L14.6015 8.90689L14.8754 9.36251L13.4353 10.203L13.8231 10.862L15.4203 9.93198L15.7105 10.4282L13.4733 11.7326L11.9115 9.05582L14.1108 7.76526L14.3954 8.26151Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 16"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M18.634 7.86581C18.5012 7.66512 18.3874 7.45351 18.2952 7.23385C18.2653 7.47248 18.2219 7.71112 18.1621 7.9444L18.032 8.53285L18.9432 8.3864L18.634 7.86581ZM20.3938 9.46325L19.594 9.51462L19.1926 8.85026L17.9371 9.04808L17.7634 9.81019L16.9879 10.008L17.8016 6.62374L18.5691 6.49915L20.3938 9.46325Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 18"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M25.3781 10.4614L24.7274 10.2065L24.9741 9.28722C24.9986 9.17873 25.0392 9.03497 25.0908 8.84789C25.1397 8.66081 25.1884 8.50345 25.2263 8.38691C25.2644 8.26481 25.3051 8.13466 25.343 7.99914L25.4216 7.77948C25.2372 8.09134 25.1667 8.2052 24.9959 8.46013C24.9254 8.5713 24.8602 8.66618 24.8061 8.74209L23.9273 9.96787L23.3605 9.82123L23.1844 8.33803C23.1709 8.25925 23.1627 8.15632 23.152 8.00719C23.1437 7.81456 23.133 7.66543 23.1303 7.56787C23.1247 7.47836 23.1247 7.35894 23.1247 7.19084L23.0462 7.56231C22.9593 7.97193 22.916 8.15076 22.7939 8.64165L22.5334 9.63684L21.8394 9.53103L22.6258 6.49658L23.6726 6.70014L23.819 8.1291C23.8543 8.46818 23.8623 8.80707 23.846 9.14614C23.9329 8.9698 24.0304 8.8017 24.1389 8.63628C24.2284 8.4954 24.3043 8.3733 24.3802 8.26481L25.1939 7.09347L26.1864 7.42967L25.3781 10.4614Z"
-            fill="currentColor"
-          />
-          <path
-            id="Fill 20"
-            fillRule="evenodd"
-            clipRule="evenodd"
-            d="M25.3042 20.4518C26.2045 19.8499 26.9449 19.5271 27.4059 19.5434C27.6798 19.5543 28.008 19.6385 28.0078 19.9774C28.0078 20.3165 27.677 20.6147 27.0316 20.8451C26.2505 21.0621 25.4343 21.1108 24.6343 20.9887C24.8486 20.7989 25.0709 20.6172 25.3042 20.4518M12.7242 21.552C12.1817 21.5139 11.9105 21.2888 11.9107 20.8631C11.9107 20.8062 11.9487 19.4476 15.0483 19.2931C15.5012 19.2661 15.954 19.2797 16.4042 19.3312C16.377 19.3366 16.0515 19.6348 16.0515 19.6348L15.8618 19.8139C15.1079 20.4945 13.8468 21.6225 12.7242 21.552M19.4605 16.4271C19.0837 16.7553 18.335 17.4656 17.6137 18.1599L16.8 18.9408C16.1412 18.8214 15.4713 18.7618 14.8016 18.7616C13.2856 18.7616 11.9107 19.1764 11.2843 19.819C11.0591 20.0224 10.9316 20.3124 10.9343 20.6162C10.9479 21.3402 11.7831 21.7119 12.038 21.8094C13.2475 22.3248 14.6414 22.1431 15.6801 21.3379C16.4013 20.7928 17.0875 20.199 17.7276 19.5591C18.2048 19.6649 18.7471 19.8114 19.3329 19.9822L19.8724 20.1368C19.1755 20.8202 18.4542 21.5361 17.7897 22.1244L17.741 22.1679C16.8188 22.9813 16.1571 23.5669 15.1972 23.7785C14.7606 23.9031 14.294 23.8705 13.8792 23.6915C13.8603 23.6944 13.8413 23.7214 13.8413 23.7214C14.3809 24.215 15.1104 24.4483 15.837 24.3587C16.7021 24.2857 17.7001 23.7461 18.7279 22.7943L20.0975 21.4385L21.0522 20.4813C21.052 20.484 23.0046 21.0536 23.0046 21.0536C23.0181 21.0563 23.0317 21.059 23.0426 21.0646C23.0181 21.0646 21.8249 22.2386 21.8249 22.2386L20.7915 23.2691C20.0973 23.9957 19.3434 24.6654 18.5407 25.2729C17.5046 26.0024 16.6342 26.3846 15.8587 26.4306C15.3515 26.4795 14.8443 26.379 14.3916 26.1458C14.3699 26.1431 14.3454 26.1703 14.3454 26.1703C14.9419 26.7831 15.7636 27.1275 16.6206 27.1222C17.8734 27.0898 18.9827 26.4879 20.6938 24.9043C21.4804 24.1747 22.2669 23.3775 22.9582 22.6618L24.2385 21.3628C25.0383 21.5474 25.8574 21.6287 26.6762 21.6099C27.8911 21.5882 28.9135 21.0051 28.9271 20.3408C28.8974 19.9477 28.6587 19.5979 28.3036 19.4296C27.856 19.1503 27.3408 19.0012 26.8121 19.0038C25.7598 19.0579 24.7619 19.4809 23.9889 20.1968L23.3435 20.7634C23.3518 20.7581 23.1076 20.6985 23.1076 20.6985L21.3232 20.202C21.8494 19.7084 22.297 19.2827 22.7037 18.9302C23.4442 18.2389 24.3824 17.7942 25.3885 17.664C25.8468 17.6343 26.6576 17.6749 26.9856 17.9109C26.9992 17.9055 27.0209 17.8729 27.0209 17.8729C26.4054 17.3035 25.5892 16.9997 24.7512 17.0267C24.1329 17.0267 23.1459 17.2218 21.9282 18.1601C21.2935 18.6916 20.6942 19.2582 20.13 19.8629C20.13 19.8549 19.2299 19.5837 19.2299 19.5837L19.0103 19.5158L18.0584 19.2283C18.0801 19.2283 18.6062 18.7184 18.6062 18.7184C19.2409 18.0949 19.8483 17.4982 20.3119 17.1051C22.0123 15.6679 23.0999 15.4781 23.8455 15.4701C24.5262 15.4321 25.2041 15.6085 25.779 15.9746C25.7926 15.9746 25.8171 15.9449 25.8171 15.9449C25.0795 15.1665 24.0435 14.7461 22.9724 14.7894C21.9203 14.7703 20.6729 15.3533 19.4605 16.4271"
-            fill="currentColor"
-          />
-        </g>
-      </g>
-    </svg>
-  );
-};
