--- src/components/svgs/index.ts
@@ -1,15 +1,0 @@
-export { EventIcon } from './EventIcon';
-export { FaqIcon } from './FaqIcon';
-export { SignUpIcon } from './SignUpIcon';
-export { KoreanAirLogo } from './KoreanAirLogo';
-export { MobileKoreanAirLogo } from './MobileKoreanAirLogo';
-export { SkyTeamLogo } from './SkyTeamLogo';
-export { SearchIcon } from './SearchIcon';
-export { MenuIcon } from './MenuIcon';
-
-export { GooglePlayDownIcon } from './GooglePlayDownIcon';
-export { AppStoreDownIcon } from './AppStoreDownIcon';
-export { YoutubeIcon } from './YoutubeIcon';
-export { InstagramIcon } from './InstagramIcon';
-export { FacebookIcon } from './FacebookIcon';
-export { TwitterIcon } from './TwitterIcon';
